Implemented Tags the right way.
	- Tag creation;
	- Tag association with messages;
	- Tag disassociation with messages;
	- Tag deletion;
	- Tag retrieval by name;
	- Tags retrieval by message.

+++ iris-core/src/test/java/br/unb/cic/iris/core/TestTagDAO.java
@@ -1,6 +1,7 @@
 package br.unb.cic.iris.core;
 
 import java.util.List;
+import java.util.Set;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -20,8 +21,7 @@ public class TestTagDAO {
 	private EmailDAO emailDao;
 	private FolderDAO folderDao;
 	
-	private EmailMessage message1;
-	private EmailMessage message2;
+	private EmailMessage message;
 	private IrisFolder folder;
 	
 	private Tag tag1;
@@ -42,12 +42,10 @@ public class TestTagDAO {
 				folder = new IrisFolder("test-folder");
 			}
 			
-			message1 = new EmailMessage("email-to-test@test.com", "email-to-test@test.com", "email-to-test@test.com", "email-to-test@test.com", "test subject 1", "test message 1");
-			message1.setFolder(folder);
-			emailDao.saveMessage(message1);
-			message2 = new EmailMessage("email-to-test@test.com", "test subject 2", "test message 2");
-			message2.setFolder(folder);
-			emailDao.saveMessage(message2);
+			message = new EmailMessage("email-to-test@test.com", "email-to-test@test.com", "email-to-test@test.com", "email-to-test@test.com", "test subject 1", "test message 1");
+			message.setFolder(folder);
+			emailDao.saveMessage(message);
+
 		}
 		catch(Exception e) {
 			throw new Exception("could not setUp the tests", e);
@@ -59,23 +57,23 @@ public class TestTagDAO {
 		try {
 			
 			tag1 = tagDao.findOrCreateByName(TAG_NAME1);
-			tag1.getMessages().add(message1);
+			tag1.getMessages().add(message);
 			tagDao.saveOrUpdate(tag1);
 			
-			List<Tag> tags = tagDao.findTagsByEmailMessage(message1);
+			List<Tag> tags = tagDao.findTagsByEmailMessage(message);
 			
-			Assert.assertTrue("The retrieved set of tags for message1 does not contain added tag1!", tags.contains(tag1));
+			Assert.assertTrue("The retrieved set of tags for message does not contain added tag1!", tags.contains(tag1));
 			
 			tag2 = tagDao.findOrCreateByName(TAG_NAME2);
-			tag2.getMessages().add(message1);
+			tag2.getMessages().add(message);
 			tagDao.saveOrUpdate(tag2);
 			
-			tags = tagDao.findTagsByEmailMessage(message1);
+			tags = tagDao.findTagsByEmailMessage(message);
 			
-			Assert.assertTrue("The retrieved set of tags for message1 does not contain added tag1!", tags.contains(tag1));
-			Assert.assertTrue("The retrieved set of tags for message1 does not contain added tag2!", tags.contains(tag2));
+			Assert.assertTrue("The retrieved set of tags for message does not contain added tag1!", tags.contains(tag1));
+			Assert.assertTrue("The retrieved set of tags for message does not contain added tag2!", tags.contains(tag2));
 			for (Tag t : tags) {
-				Assert.assertTrue("Tag " + t.getName() + " should have message1 but doens't!", t.getMessages().contains(message1));
+				Assert.assertTrue("Tag " + t.getName() + " should have message1 but doens't!", t.getMessages().contains(message));
 			}
 			
 		}
@@ -84,30 +82,47 @@ public class TestTagDAO {
 		}
 	}
 	
-//	@Test
-//	public void removeTagFromMessage() throws Exception {
-//		tag1 = tagDao.findOrCreateByName(TAG_NAME1);
-//		tag1.getMessages().add(message2);
-//		tagDao.saveOrUpdate(tag1);
-//		
-//		List<Tag> tags = tagDao.findTagsByEmailMessage(message2);
-//		
-//		Assert.assertTrue("The retrieved set of tags for message2 does not contain added tag1!", tags.contains(tag1));
-//		
-//		tag1 = tagDao.findOrCreateByName(TAG_NAME1);
-//		tag1.getMessages().remove(message2);
-//		tagDao.saveOrUpdate(tag1);
-//		
-//		Assert.assertTrue("Failed to remove tag1 from message2!", !tags.contains(tag1));
-//	}
+	@Test
+	public void removeTagFromMessage() throws Exception {
+		tag1 = tagDao.findOrCreateByName(TAG_NAME1);
+		tag1.getMessages().add(message);
+		tagDao.saveOrUpdate(tag1);
+		
+		List<Tag> tags = tagDao.findTagsByEmailMessage(message);
+		
+		Assert.assertTrue("The retrieved set of tags for message does not contain added tag1!", tags.contains(tag1));
+		
+		tag1 = tagDao.findOrCreateByName(TAG_NAME1);
+		Set<EmailMessage> msgs = tag1.getMessages();
+		tag1.getMessages().remove(message);
+		tagDao.saveOrUpdate(tag1);
+		
+		tags = tagDao.findTagsByEmailMessage(message);
+		
+		Assert.assertTrue("Failed to remove tag1 from message!", !tags.contains(tag1));
+	}
+	
+	@Test
+	public void deleteTag() throws Exception {
+		tag1 = tagDao.findOrCreateByName(TAG_NAME1);
+		tag1.getMessages().add(message);
+		tagDao.saveOrUpdate(tag1);
+		
+		tagDao.delete(tag1);
+		List<Tag> tags = tagDao.findTagsByEmailMessage(message);
+		Assert.assertTrue("Failed to delete tag1!", !tags.contains(tag1));
+		
+		tag1 = null;
+	}
 	
 	@After
 	public void tearDown() throws Exception {
 		try {
-			tagDao.delete(tag1);
-			tagDao.delete(tag2);
-			emailDao.delete(message1);
-			emailDao.delete(message2);
+			if (tag1 != null)
+				tagDao.delete(tag1);
+			if (tag2 != null)
+				tagDao.delete(tag2);
+			emailDao.delete(message);
 			folderDao.delete(folder);
 		}
 		catch(Exception e) {
